created NonMatchingMockWallet
