+++ dash/src/test_utils/mod.rs
@@ -1,2 +1,3 @@
+mod block;
 mod blockhash;
 mod chainlock;
